Container component in popoverRef has been removed. PopoverContainer now implements what the container for popoverRef. popoverRef should work and behave in the same way as before. Keeping the import and export off popoveplacement and justification in the popoverref to not cause any breaking changes. Old Popover in general has been removed and replaced with popoverContainer. popover was currently not exported or in use. PopoverContainer is copypaste of the container from popoverRef. The component itself is fairly barebones and only really displays a div with content is supplied. In almost all cases one should use popoverRef instead of popoverContainer directly.
