--- src/components/data/HangingGarden/styles.less
@@ -1,10 +1,0 @@
-.container {
-  width: 100%;
-  height: 100%;
-  overflow: auto;
-  position: relative;
-}
-
-.popoverContainer {
-  position: absolute;
-}
