--- src/components/general/Popover/positioning.js
@@ -1,9 +1,0 @@
-export const popoverHorizontalPositions = {
-    left: 'left',
-    right: 'right',
-};
-
-export const popoverVerticalPositions = {
-    top: 'top',
-    bottom: 'bottom',
-};
