--- src/components/general/Popover/styles.less
@@ -1,73 +1,0 @@
-@import "~styles/variables.less";
-
-.container {
-    position: relative;
-    pointer-events: all;
-    max-height: 75vh;
-    display: flex;
-
-    &.isRelative {
-        position: absolute;
-        top: 100%;
-    }
-
-    &.left {
-        left: 0;
-
-        .arrow {
-            left: 11px;
-        }
-    }
-
-    &.right {
-        right: 0;
-
-        .arrow {
-            right: 11px;
-        }
-    }
-
-    &.isInset {
-        top: calc(100% - calc(var(--grid-unit) * 2px));
-
-        @media (max-width: var(--mobile-max-width)) {
-            top:100%;
-        }
-
-        &.left {
-            left: calc(var(--grid-unit) * 3px);
-        }
-
-        &.right {
-            right: calc(var(--grid-unit) * 1px);
-        }
-    }
-
-    .content {
-        background: white;
-        overflow: hidden;
-        border-radius: 4px;
-        box-shadow: 0 0 4px rgba(0, 0, 0, .1), 0 4px 4px rgba(0, 0, 0, .05);
-        display: flex;
-        max-height: 100%;
-        box-sizing: border-box;
-        padding: calc(var(--grid-unit) * 1px)
-    }
-
-    .arrow {
-        position: absolute;
-        bottom: 100%;
-        width: 10px;
-        height: 4px;
-
-        > svg {
-            position: absolute;
-            top: -4px;
-            left: -4px;
-
-            path {
-                box-shadow: 0 0 4px rgba(0, 0, 0, 0.1), 0 4px 4px rgba(0, 0, 0, 0.05);
-            }
-        }
-    }
-}
