+++ src/components/general/PopoverContainer/Arrow.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from '../styles.less';
+import styles from './styles.less';
 
 export default () => {
     return (
