+++ src/hooks/usePopoverRef/styles.less
@@ -1,104 +1,3 @@
 .container {
     position: absolute;
 }
-
-.popoverContainer {
-    --arrow-width: 12px;
-    --arrow-height: 8px;
-    
-    padding: calc(var(--grid-unit) * 2px);
-    background: white;
-    border: 1px solid var(--color-black-alt4);
-    border-radius: 4px;
-    position: absolute;
-    pointer-events: all;
-
-    &.fillWithContent {
-        padding: 0;
-    }
-
-    .arrow {
-        position: absolute;
-    }
-
-    &.below {
-        top: 100%;
-        margin-top: calc(var(--grid-unit) * 1px);
-
-        .arrow {
-            bottom: 100%;
-        }
-    }
-
-    &.above {
-        bottom: 100%;
-        margin-bottom: calc(var(--grid-unit) * 1px);
-
-        .arrow {
-            top: 100%;
-            transform: rotate(180deg);
-        }
-    }
-
-    &.left {
-        right: 100%;
-        margin-right: calc(var(--grid-unit) * 1px);
-
-        .arrow {
-            transform: rotate(90deg);
-
-            left: 100%;
-            margin-left: -2px;
-        }
-    }
-
-    &.right {
-        left: 100%;
-        margin-left: calc(var(--grid-unit) * 1px);
-
-        .arrow {
-            transform: rotate(-90deg);
-
-            right: 100%;
-            margin-right: -2px;
-        }
-    }
-
-    &.start.below, &.start.above {
-        left: 0;
-
-        .arrow {
-            left: calc(var(--grid-unit) * 1px);
-        }
-
-        &.isCentered {
-            left: calc(50% - (var(--grid-unit) * 1px) - (var(--arrow-width) / 2));
-        }
-    }
-
-    &.center.below, &.center.above {
-        left: 50%;
-        transform: translateX(-50%);
-
-        .arrow {
-            left: 50%;
-            transform: translateX(-50%);
-        }
-
-        &.above .arrow {
-            transform: translateX(-50%) rotate(180deg);
-        }
-    }
-
-    &.end.below, &.end.above {
-        right: 0;
-
-        .arrow {
-            right: calc(var(--grid-unit) * 1px);
-        }
-
-        &.isCentered {
-            right: calc(50% - (var(--grid-unit) * 1px) - (var(--arrow-width) / 2));
-        }
-    }
-}
